Created a new Class "Hud.java" in the package "scenes".
The Hud.java will contain and display our overlay information ingame,
such as Health, Score, Time ++ whatever we like.

--------

In Hud.java::26 the viewport Width and Height is set to be the size of
our gamescreen.
If we later want to have a camera moving, on a map size bigger than our
camera, the Width and Height need to be adjusted to the size of the
camera.
We can easily add more Labels by repeating the sequence of

...Label = new Label(...)

--------

Changes to create the Hud on our screen and display it:

Game.java::19 added hud private variable.
Game.java::39 addded hud = new Hud(game.batch);

Game.java::53 & 54
        batch.setProjectionMatrix(hud.stage.getCamera().combined);
        hud.stage.draw();

--------

Coin - Test Object

Objects to gain score can be anything, Coin can be changed to something
more suitable later, dependent on our Theme in the game.

Added a new Entity Coin.java, and added under EntityType::6.

Coin is one of the entities that are collidable for Entity.Player and
the outcome will be to get +100 Score.

--------

New package: Tools

Creates a new class inside "Tools" with the name "CollisionRect.java"

CollisionRect.java will work as an hitbox for the game, where it will
detect collisions between entities.

Source: https://www.youtube.com/watch?v=xMxtf0sJvr4&ab_channel=HollowBit


Adds a CollisionRect to Entity, with the fieldvariable in Entity::14,
constructs in ::20, and the rectangle will move accordingly to other
entities at Entity.moveX::46 and Entity.moveY::52


In GameMap.java, i added player as a variable itself, and ArrayList for
coins (can probably just be CollidableEntities or similar in future, or
every different entity gets a different list.)
. in the render function ::26, i've made a new Array for the coins to be
removed, and coins the list will remove overlapping objects after the
forloop.

after removing coins with collisions, the remainings will be rendered.

--------

Adding score to the HUD inside Game.java when a coin is removed have
been difficult, since the hud seems difficult to access.
Might can use other ways to track score/display HUD's in the future that
suits our project build better.

Picture goldCoin1.png has a bad aspect ratio compared to the game, so
the hitbox appaers bigger than it is. This can be fixed in the future
with entities with better aspect ratios.