Implemented that players and other active entities can die
Hp for skeleton set to 1, player hp set to 2.

If an active entity goes under the stage (-y coordinate) the game sets
their health to 0 and kills them in a new loop in gameMap that removes
all dead entities. Current issue with implementation: Players' scores
are removed on death. Possible fix: the scores can be stored somewhere
outside the player class